refactor: 크루 생성 페이지 상태, 로직 분리
