+++ src/pages/CreateCrewPage/CreateCrewPage.tsx
@@ -1,6 +1,6 @@
-import { useState } from 'react';
-import { FormProvider, useForm } from 'react-hook-form';
-import { useNavigate } from 'react-router-dom';
+import { FormProvider } from 'react-hook-form';
+
+import { MAX_MEMBER_COUNT_LIST } from '@pages/CreateCrewPage/consts/createCrewOptions';
 
 import { ConditionalFormInput } from '@components/ConditionalFormInput';
 import { Header } from '@components/Header';
@@ -8,21 +8,13 @@ import { Modal } from '@components/Modal';
 import { TextArea } from '@components/TextArea';
 import { Button } from '@components/shared/Button';
 import { Text } from '@components/shared/Text';
-import { useToggleButtons } from '@components/shared/ToggleButton';
 import { VirtualScroll } from '@components/shared/VirtualScroll';
 
-import { useCrewMutation } from '@hooks/mutations/useCrewMutation';
 import { useHeaderTitle } from '@hooks/useHeaderTitle';
 
 import { theme } from '@styles/theme';
 
-import { useLoginInfoStore } from '@stores/loginInfo.store';
-
-import { PostCrewRequest } from '@type/api/crews';
-
-import { MAX_MEMBER_COUNT_LIST } from '@consts/createCrewOptions';
 import { SEOUL } from '@consts/location';
-import { PATH_NAME } from '@consts/pathName';
 
 import {
   PageLayout,
@@ -30,67 +22,37 @@ import {
   ScrollBox,
   StyledCreateForm,
   StyledEmptyContainer,
+  StyledModalContent,
   StyledModalHeader,
   StyledSelectBox,
   StyledSelectedLocationButton,
   StyledTitle,
   StyledToggleButton,
 } from './CreateCrewPage.styles';
+import { useCreateCrewPage } from './useCreateCrewPage';
 
 export const CreateCrewPage = () => {
-  const loginInfo = useLoginInfoStore((state) => state.loginInfo);
-
-  if (!loginInfo?.id) {
-    throw new Error('로그인이 필요한 서비스입니다.');
-  }
-  const navigate = useNavigate();
-
-  const { mutate } = useCrewMutation();
-  const methods = useForm();
-
   const { entryRef, showHeaderTitle } = useHeaderTitle<HTMLDivElement>();
-
-  const [name, setName] = useState<string>('');
-  const [content, setContent] = useState<string>('');
-  const [maxMemberCount, setMaxMemberCount] = useState<string>('');
-  const [selectedLocation, setSelectedLocation] = useState<string[]>();
-
-  const [isOpenMaxMemberCountModal, setIsOpenMaxMemberCountModal] =
-    useState(false);
-  const [isOpenAddressDepth2Modal, setIsOpenAddressDepth2Modal] =
-    useState(false);
+  const {
+    state,
+    methods,
+    onSubmit,
+    setName,
+    setContent,
+    setMaxMemberCount,
+    handleToggleLocation,
+    toggleMaxMemberCountModal,
+    toggleAddressDepth2Modal,
+  } = useCreateCrewPage();
 
   const {
-    handleToggle: handleToggleLocation,
-    selectedItems: selectedLocations,
-  } = useToggleButtons({
-    onToggle: setSelectedLocation,
-    isMultipleSelect: false,
-  });
-
-  const toggleMaxMemberCountModal = () => {
-    setIsOpenMaxMemberCountModal((prev) => !prev);
-  };
-
-  const toggleAddressDepth2Modal = () => {
-    setIsOpenAddressDepth2Modal((prev) => !prev);
-  };
-
-  const onSubmit = () => {
-    const data: PostCrewRequest = {
-      name,
-      content,
-      maxMemberCount: parseInt(maxMemberCount),
-      addressDepth1: '서울시',
-      addressDepth2: selectedLocation![0],
-    };
-
-    mutate(data, {
-      onSuccess: ({ crewId }) => {
-        navigate(PATH_NAME.GET_CREWS_PATH(String(crewId)));
-      },
-    });
-  };
+    name,
+    maxMemberCount,
+    selectedLocation,
+    selectedLocations,
+    isOpenMaxMemberCountModal,
+    isOpenAddressDepth2Modal,
+  } = state;
 
   return (
     <PageLayout>
@@ -155,13 +117,7 @@ export const CreateCrewPage = () => {
                   주로 활동하는 지역을 선택해 주세요!
                 </Text>
               </StyledModalHeader>
-              <Modal.Content
-                style={{
-                  display: 'flex',
-                  justifyContent: 'center',
-                  alignContent: 'center',
-                }}
-              >
+              <StyledModalContent>
                 <ScrollBox>
                   <StyledSelectBox border="none">
                     {Object.values(SEOUL).map((location) => (
@@ -175,7 +131,7 @@ export const CreateCrewPage = () => {
                     ))}
                   </StyledSelectBox>
                 </ScrollBox>
-              </Modal.Content>
+              </StyledModalContent>
             </Modal>
             <TextArea
               title="상세설명을 입력해 주세요!"
